feat: adds capture method
